NN-3171 improvements
