+++ helm_deploy/values-dev.yaml
@@ -21,5 +21,5 @@ env:
   AUTH_URI_ROOT: "https://sign-in-dev.hmpps.service.justice.gov.uk/auth"
   JAVA_OPTS: "-Xmx512m"
   SPRING_SECURITY_OAUTH2_RESOURCESERVER_JWT_JWK_SET_URI: "https://sign-in-dev.hmpps.service.justice.gov.uk/auth/.well-known/jwks.json"
-  WOMENS_ESTATE: "MDI,LEI"
+  PRISONS_WITH_OFFENDERS_THAT_HAVE_COMPLEX_NEEDS: "MDI,LEI"
   COMPLEXITY_OF_NEED_URI: "https://hmpps-complexity-of-need-staging.apps.live-1.cloud-platform.service.justice.gov.uk"
