+++ helm_deploy/values-prod.yaml
@@ -21,8 +21,6 @@ env:
   AUTH_URI_ROOT: "https://sign-in.hmpps.service.justice.gov.uk/auth"
   JAVA_OPTS: "-Xmx512m"
   SPRING_SECURITY_OAUTH2_RESOURCESERVER_JWT_JWK_SET_URI: "https://sign-in.hmpps.service.justice.gov.uk/auth/.well-known/jwks.json"
-  COMPLEXITY_OF_NEED_URI: "https://hmpps-complexity-of-need-staging.apps.live-1.cloud-platform.service.justice.gov.uk"
-
 whitelist:
   office: "217.33.148.210/32"
   health-kick: "35.177.252.195/32"
