+++ src/main/java/uk/gov/justice/digital/hmpps/keyworker/services/ComplexityOfNeedService.kt
@@ -3,21 +3,37 @@ package uk.gov.justice.digital.hmpps.keyworker.services
 import com.microsoft.applicationinsights.TelemetryClient
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.keyworker.events.ComplexityOfNeedLevel
 
+interface ComplexityOfNeed {
+  fun removeOffendersWithHighComplexityOfNeed(prisonId: String, offenders: Set<String>): Set<String>
+  fun onComplexityChange(offenderNo: String, level: ComplexityOfNeedLevel)
+}
+
 @Service
+@ConditionalOnExpression("T(org.apache.commons.lang3.StringUtils).isBlank('\${complexity_of_need_uri}')")
+class StubComplexityOfNeed : ComplexityOfNeed {
+  override fun removeOffendersWithHighComplexityOfNeed(prisonId: String, offenders: Set<String>): Set<String> =
+    offenders
+
+  override fun onComplexityChange(offenderNo: String, level: ComplexityOfNeedLevel) {}
+}
+
+@Service
+@ConditionalOnExpression("T(org.apache.commons.lang3.StringUtils).isNotBlank('\${complexity_of_need_uri}')")
 class ComplexityOfNeedService(
   val keyworkerService: KeyworkerService,
   val complexityOfNeedGateway: ComplexityOfNeedGateway,
-  @Value("\${womens.estate}") val prisonsWithOffenderComplexityNeeds: Set<String>,
+  @Value("\${prisons.with.offenders.that.have.complex.needs}") val prisonsWithOffenderComplexityNeeds: Set<String>,
   val telemetryClient: TelemetryClient
-) {
+) : ComplexityOfNeed {
   companion object {
     private val log = LoggerFactory.getLogger(this::class.java)
   }
 
-  fun removeOffendersWithHighComplexityOfNeed(prisonId: String, offenders: Set<String>): Set<String> {
+  override fun removeOffendersWithHighComplexityOfNeed(prisonId: String, offenders: Set<String>): Set<String> {
     if (!prisonsWithOffenderComplexityNeeds.contains(prisonId)) return offenders
 
     val complexOffenders = complexityOfNeedGateway
@@ -33,7 +49,7 @@ class ComplexityOfNeedService(
     }.toSet()
   }
 
-  fun onComplexityChange(offenderNo: String, level: ComplexityOfNeedLevel) {
+  override fun onComplexityChange(offenderNo: String, level: ComplexityOfNeedLevel) {
     telemetryClient.trackEvent(
       "complexity-of-need-change",
       mapOf(
