+++ src/main/resources/application.properties
@@ -34,4 +34,5 @@ application.insights.ikey=
 logging.uris.exclude.regex=.*/health|.*/health/ping|.*/info|.*/ping|.*/swagger.*|/webjars/.*|.*/v2/api-docs
 graceful.shutdown.enabled=true
 
-womens.estate=
+
+prisons.with.offenders.that.have.complex.needs=
